Remove ClayTDM and Spookghetto
Spookghetto does not play well. Tons of random soul sand, cobwebs, and the center is a pain to boat through if you don't want to be shot or tapped with the axe. It's one of the few one hit kill maps left, and it's not the funnest of them all to keep. It's a pain to navigate.

ClayTDM is too flat and was removed a while ago. The gameplay is stale and it's not very fun.